Support building with Zig (experimental)
PR #598 (bdwgc).

This adds a build.zig that allows building the garbage collector using
the Zig build system.  It implements a subset of the configuration
options offered by CMake and following the naming convention in CMake
for all those options.

The README is slightly updated adding an explanation of zig.

The build.zig follows the format used in zig v0.12.0-dev.1814.
The final 0.12 version is not out yet, so users will have to use
a build from the zig master branch for now.  This is not ideal but the
whole zig build system is so new that there have been backwards
incompatible changes from 0.11 which is why we opt to use 0.12 that is
likely going to be closer to the final form.

* README.md (Building and Installing): Add subsection about Zig.
* Makefile.am (EXTRA_DIST): Add build.zig item.
* build.zig: New file.

Co-authored-by: Ivan Maidanski <ivmai@mail.ru>