Fix undefined behavior caused by potential overflow in MAKE_COOLER
Issue #612 (bdwgc).

Check for a potential overflow before the addition (or subtraction)
operation.

This commit also changes the behavior of MAKE_COOLER() in case of the
overflow - it retains the pointer unchanged instead of setting it to
zero or GC_WORD_MAX.  This change should not affect the logic of
GC_clear_stack() which uses MAKE_COOLER().

* include/private/gc_priv.h (MAKE_COOLER): Rename x,y arguments to p,d;
do not perform addition/subtraction if it may lead to overflow; do
nothing to p if overflow is expected.