Fix per_object_helper() and Barrett's diagram after changing hb_sz units (fix of commit e35a4171f)
Now hb_sz is the size of the block in bytes, not words.

* backgraph.c (per_object_helper): Replace BYTES_TO_WORDS(HBLKSIZE) to
HBLKSIZE.
* docs/tree.md (A picture): Change order of hb_sz, hb_descr, hb_map,
fields of hblkhdr (to match that in gc_priv.h); mention hb_n_marks
field in hblkhdr; change type of hb_sz to size_t; do not mention
"words" for hb_sz and MARK_BITS_SZ; specify type for hb_marks field in
hblkhdr and reformat its comment.