Do not mix system and bdwgc headers in same include group
Issue #364 (bdwgc).

This is needed to prevent clang-format to change the order of included
header files where the order is important.

* cord/tests/de.c [WIN32]: Move include de_win.h down to be in
a separate include group (from that of windows.h).
* extra/AmigaOS.c: Place a blank line after include gc_priv.h.
* include/private/gc_priv.h [ANY_MSWIN]: Place a blank line between
include windows.h and winbase.h (and add a comment).
* pthread_stop_world.c [GC_ENABLE_SUSPEND_THREAD]: Place a blank line
between include sys/time.h and javaxfc.h; reformat comment.
* tests/gctest.c: Place a blank line between include gc.h and
javaxfc.h.
* tests/weakmap.c [GC_PTHREADS]: Move include errno.h and pthread.h
upper (to be in a separate include group).