Remove implicit GC initialization by GC_set_warn_proc on Win32
In case of multi-threaded libgc build for Cygwin, we do not check that
the collector is initialized when GC_set_warn_proc is invoked.
For other Win32 targets (in case of multi-threaded build), we do not
perform collector initialization by GC_set_warn_proc.

* misc.c [GC_WIN32_THREADS && CYGWIN32] (GC_set_warn_proc): Do not
assert that GC is initialized.
* misc.c [GC_WIN32_THREADS && !CYGWIN32] (GC_set_warn_proc): Do not
call GC_init().