Export CMake libraries with aliases
PR #675 (bdwgc).

This allows bdwgc to be included in a cmake build using
`add_subdirectory`, `FetchContent`, or similar.  It then relies on the
CMake CMP0028 policy to ensure the library is found in any
`target_link_libraries` usage.

Without the alias, `target_link_libraries` will not be able to find
`gc`, `gccpp`, etc.

* CMakeLists.txt (add_library): Specify alias for gc.
* CMakeLists.txt [enable_cplusplus] (add_library): Specify alias for
gccpp.
* CMakeLists.txt [enable_cplusplus && enable_throw_bad_alloc_library]
(add_library): Specify alias for gctba.
* CMakeLists.txt [build_cord] (add_library): Specify alias for cord.