Fix mark lock in child process after fork
PR #678 (bdwgc).

Reinitialize the mark lock in `fork_child_proc`.  The reason is the
same as for mutex-based GC_allocate_ml.  It is not legal to reuse any
locks in child processes, at least, on Haiku; and the POSIX
specification says that unlocking a mutex that a thread does not own
is an undefined behavior anyway.

* pthread_support.c [CAN_HANDLE_FORK && PARALLEL_MARK
&& !GC_WIN32_THREADS] (fork_child_proc): Call `pthread_mutex_destroy()`
for `mark_mutex` ignoring return value and then call
`pthread_mutex_init()` for `mark_mutex`; add comment.

Co-authored-by: Ivan Maidanski <ivmai@mail.ru>