Limit number of GC_collect_or_expand calls in GC_allocobj/large
Only a few iterations are expected at most in `GC_allocobj` and
`GC_alloc_large`, otherwise something is wrong in one of the functions
called inside the corresponding loop.

* alloc.c (MAX_ALLOCOBJ_RETRIES): Define macro.
* malloc.c (MAX_ALLOCLARGE_RETRIES): Likewise.
* alloc.c (GC_allocobj): Replace `retry` local variable to `retry_cnt`
(of `int` type); increment `retry_cnt` at the end of the loop; add
comment.
* malloc.c (GC_alloc_large): Likewise.
* alloc.c (GC_allocobj): Call `ABORT()` if `retry_cnt` is greater than
`MAX_ALLOCOBJ_RETRIES` (at the beginning of the loop).
* malloc.c (GC_alloc_large): Call `ABORT()` if `retry_cnt` is greater
than `MAX_ALLOCLARGE_RETRIES` (at the beginning of the loop).