Avoid abort on ResumeThread failure if Win32 DllMain closes thread handle
PR #712 (bdwgc).

It is possible that `DllMain()` has set the thread handle to NULL since
the `SuspendThread()` call and that is why `ResumeThread()` has failed.
In this case, `GC_suspend()` and `GC_start_world()` should not abort.
Same applies to `GetThreadContext()` call inside `GC_push_stack_for()`.

It might be the scenario like this:
1. `GC_suspend` calls `SuspendThread` on a valid handle;
2. Within the `SuspendThread` call a context switch occurs to `DllMain`
   (before the thread has actually been suspended);
3. `DllMain` sets `t->handle` to NULL, but does not yet close the
   handle;
4. A context switch occurs returning to `SuspendThread` which completes
on the handle that was originally passed into it;
5. Then `ResumeThread` attempts to run on t->handle which is now NULL.

* win32_threads.c [!GC_NO_THREADS_DISCOVERY] (GC_suspend,
GC_start_world): Call `WARN()` and return from the function instead of
`ABORT()` if `ResumeThread()` has failed but `t->handle` after it
(fetched atomically with a barrier) is NULL; add FIXME item; add
comment.
* win32_threads.c [GC_NO_THREADS_DISCOVERY] (GC_start_world): Do not
check `GC_win32_dll_threads` variable (assume it is always false).
* win32_threads.c [!GC_NO_THREADS_DISCOVERY] (GC_start_world): Replace
`THREAD_HANDLE(p)` to `p->handle` for code where `GC_win32_dll_threads`
is true.
* win32_threads.c [!GC_NO_THREADS_DISCOVERY] (GC_push_stack_for):
Return zero from the function instead of `ABORT()` if
`GetThreadContext()` has failed but `thread->handle` after it (fetched
atomically with a barrier) is NULL.

Co-authored-by: Ivan Maidanski <ivmai@mail.ru>