Add option to disable output entirely
This is mostly useful in environments where printf adds a non-negligible
amount of code (for example, WebAssembly in the browser). Disabling all
output shrinks the binary. Of course, this option should not be used
when testing/debugging the GC.

On WebAssembly, the reduction in binary size is 27.5kB. On Linux (arm64)
statically linked against musl the difference is around 16kB in my
setup. I expect this is mostly because printf and related functions do
not need to be linked in for simple programs that don't need these.