Add option to disable output entirely
This is mostly useful in environments where printf adds a non-negligible
amount of code (for example, WebAssembly in the browser). Disabling all
output shrinks the binary. Of course, this option should not be used
when testing/debugging the GC.

I didn't test the binary size difference on WebAssembly, but on Linux
(arm64) with a statically linked musl libc the difference is 20kB in my
setup.